requirements.txt
gunicorn
flask
flask_cors
pandas
numpy
pillow
firebase_admin